requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.2
distro==1.8.0
dnspython==2.4.2
environs==9.5.0
exceptiongroup==1.2.0
fastapi==0.104.1
Flask==3.0.0
frozenlist==1.4.0
greenlet==3.0.1
grpcio==1.58.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.340
langsmith==0.0.66
loguru==0.7.2
MarkupSafe==2.1.3
marshmallow==3.20.1
minio==7.2.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.5
packaging==23.2
pandas==2.1.3
pinecone-client==2.2.4
protobuf==4.25.1
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
pymilvus==2.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
slack-bolt==1.18.1
slack-sdk==3.26.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.18
Werkzeug==3.0.1
yarl==1.9.3
